Drop 0.6 support and simplify package structure
