+++ REQUIRE
@@ -1,4 +1,4 @@
-julia 0.6.0
+julia 0.7.0-DEV.602
 AbstractFFTs 0.2.0
 Reexport
 Compat 0.27.0
